pyproject.toml
[tool.poetry]
name = "my-automation"
version = "0.1.0"
description = ""
authors = ["uuboyscy <aegis12321@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
bs4 = "^0.0.2"
yfinance = "^0.2.44"
pendulum = "^3.0.0"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
